// src/generated/models/Cr9cd_contact_beneficiariesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_contact_beneficiariescr9cd_future_priority_flag = {
  100000000: 'Elevated',
  100000001: 'Normal',
  100000002: 'Deprioritized'
} as const;
export type Cr9cd_contact_beneficiariescr9cd_future_priority_flag = keyof typeof Cr9cd_contact_beneficiariescr9cd_future_priority_flag;
export const Cr9cd_contact_beneficiariescr9cd_type = {
  100000000: 'Customer',
  100000001: 'Employee'
} as const;
export type Cr9cd_contact_beneficiariescr9cd_type = keyof typeof Cr9cd_contact_beneficiariescr9cd_type;
export const Cr9cd_contact_beneficiariescr9cd_value_tier = {
  100000000: 'Platinum',
  100000001: 'Gold',
  100000002: 'Silver',
  100000003: 'Bronze',
  100000004: 'Prospect'
} as const;
export type Cr9cd_contact_beneficiariescr9cd_value_tier = keyof typeof Cr9cd_contact_beneficiariescr9cd_value_tier;
export const Cr9cd_contact_beneficiariesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_contact_beneficiariesstatecode = keyof typeof Cr9cd_contact_beneficiariesstatecode;
export const Cr9cd_contact_beneficiariesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_contact_beneficiariesstatuscode = keyof typeof Cr9cd_contact_beneficiariesstatuscode;

export interface Cr9cd_contact_beneficiariesBase {
  cr9cd_attended_count?: number;
  cr9cd_awarded_count?: number;
  cr9cd_company?: string;
  cr9cd_contact_beneficiaryid: string;
  "cr9cd_Crm_Account@odata.bind"?: string;
  "cr9cd_Crm_Contact@odata.bind"?: string;
  cr9cd_directory_user_id?: string;
  cr9cd_email?: string;
  cr9cd_future_priority_flag?: Cr9cd_contact_beneficiariescr9cd_future_priority_flag;
  cr9cd_last_ticket_date?: string;
  cr9cd_lifetime_business_generated?: number;
  cr9cd_name?: string;
  cr9cd_no_show_count?: number;
  cr9cd_notes?: string;
  cr9cd_phone?: string;
  cr9cd_title?: string;
  cr9cd_type?: Cr9cd_contact_beneficiariescr9cd_type;
  cr9cd_value_tier?: Cr9cd_contact_beneficiariescr9cd_value_tier;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_contact_beneficiariesstatecode;
  statuscode?: Cr9cd_contact_beneficiariesstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_contact_beneficiaries extends Cr9cd_contact_beneficiariesBase {
  cr9cd_crm_accountname?: string;
  cr9cd_crm_accountyominame?: string;
  cr9cd_crm_contactname?: string;
  cr9cd_crm_contactyominame?: string;
  cr9cd_future_priority_flagname?: string;
  cr9cd_lifetime_business_generated_base?: number;
  cr9cd_typename?: string;
  cr9cd_value_tiername?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  cr9cd_crm_account?: object;
  _cr9cd_crm_account_value?: string;
  cr9cd_crm_contact?: object;
  _cr9cd_crm_contact_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}